requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
django-jwt-auth==0.0.2
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
